application/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pytest-mock==3.11.1
python-dotenv==1.0.0
requests==2.31.0